Use deferred imports for the main App code, to improve chunking.
This splits the JS in two parts, one with the minimal JS required to
bootstrap a flutter app, and another with my actual app code and its
dependencies.
