+++ lib/main.dart
@@ -1,35 +1,13 @@
 import 'package:flutter/material.dart';
 
-import 'src/home_page.dart';
-import 'src/restaurant_page.dart';
+import 'src/app.dart' deferred as app;
 
-void main() => runApp(FriendlyEatsApp());
-
-class FriendlyEatsApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'FriendlyEats',
-      onGenerateRoute: (settings) {
-        switch (settings.name) {
-          case FriendlyEatsRestaurantPage.route:
-            final FriendlyEatsRestaurantPageArguments arguments =
-                settings.arguments;
-            return MaterialPageRoute(
-                builder: (context) => FriendlyEatsRestaurantPage(
-                      restaurantId: arguments.id,
-                    ));
-            break;
-          default:
-            // return MaterialPageRoute(
-            //     builder: (context) => FriendlyEatsRestaurantPage(
-            //           restaurantId: 'lV81npEeboEActMpUJjn',
-            //         ));
-            // Everything defaults to home, but maybe we want a custom 404 here
-            return MaterialPageRoute(
-                builder: (context) => FriendlyEatsHomePage());
-        }
-      },
-    );
-  }
+void main() {
+  final Future<void> loadedLibrary = app.loadLibrary();
+  runApp(
+    FutureBuilder(
+      future: loadedLibrary,
+      builder: (snapshot, context) => app.FriendlyEatsApp(),
+    ),
+  );
 }
